Add controls to activate/deactivate mappings
